requirements.txt
selenium
webdriver-manager
selenium-wire
blinker==1.7.0
beautifulsoup4
pandas
openpyxl
groq
torch 
torchvision 
torchaudio
marker-pdf
json-repair
megaparse
python-magic-bin